Afteryoon

+++ src/main/java/com/app/ycommerce/dto/ProductDto.java
@@ -9,14 +9,14 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ProductDto {
 
-	private int productId;
-	private int categoryId;
+	private Long productId;
+	private Long categoryId;
 	private String productName;
 	private String productDescription;
 	private int price;
 	private int inventory;
 
-	public ProductDto(int productId, int categoryId, String productName, String productDescription, int price,
+	public ProductDto(Long productId, Long categoryId, String productName, String productDescription, int price,
 		int inventory) {
 		this.productId = productId;
 		this.categoryId = categoryId;
